create form component, clean up pages dir
